Fixed a compilation error for clang and noted a potential bug.
The compilation error was due to a warning about extra braces needed
around an initializer list. This warning occurs only on clang.

The potential bug is related to the discrete weighted distribution used
for kmeans++ initialization. I suspect in the current version this code
may not perform as indentended when distances are small.